requirements_old.txt
connectorx
IPython
pandas
plotly
polars
polars-ta
psycopg2-binary
pyarrow
python-binance
python-decouple
PyYAML
quantstats
requests
sqlalchemy
websocket-client